Moved taper-configuration route into PatientPage
